Safari-609.1.11.1/LayoutTests/fast/loader/frames-with-unload-handlers-in-page-cache-expected.txt

frame "<!--frame2-->" - has 1 onunload handler(s)
ALERT: Subframe onload
ALERT: Subframe onpageshow
CONSOLE MESSAGE: line 16: Use of window.alert is not allowed while unloading a page.
CONSOLE MESSAGE: line 18: Use of window.alert is not allowed while unloading a page.
CONSOLE MESSAGE: line 10: Use of window.alert is not allowed while unloading a page.
